requirements.txt
jupyterlab
scikit-learn
pandas
matplotlib
expand-folders
maestro @ git+https://github.com/jodafons/maestro.git@main
dorothy @ git+https://github.com/bricstb-softwares/dorothy.git@master
pybeamer
atlas_mpl_style
tqdm
pandarallel
loguru
tensorflow-gpu==2.10.0
opencv-python
